requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
requests
httpx
supabase
pyjwt
starlette
google-generativeai
google-auth
Pillow
python-multipart
sqlalchemy[asyncio]
asyncpg
pydantic
itsdangerous